+++ src/bot/index.ts
@@ -15,7 +15,6 @@ const client = new Client({
   intents: [
     GatewayIntentBits.GuildMessages,
     GatewayIntentBits.Guilds,
-    GatewayIntentBits.MessageContent,
   ],
 });